Add direct pipe to muxrpc api in main thread
Pull-streams in the renderer thread were taking up too much of the
process without calling back, causing the UI to hang. Instead passing
stream to muxrpc api running in main thread. Requires pauls-electron-rpc
with duplex support, package used for that is mikes-electron-rpc, but
that package will be discontinued if my PR is accepted.